[ADD] login modal

+++ src/features/TimelineCard/TimelineCard.js
@@ -11,6 +11,8 @@ import { useTheme } from '../../shared/context/ThemeContext'
 import { ROUTE } from '../../shared/constants/NavigationConstants'
 import { Swiper } from '../../shared/components/Swiper'
 import { checkErr } from '../../utils/CommonUtils'
+import { LoginModalComponent } from '../../shared/components/LoginModal'
+import { useSelector } from 'react-redux'
 
 export const TimelineCard = ({ avatar, name, place, caption, links, time, date, comments, feedId, handleComment, postLikes, setRefresh, accId, postAccId, handleClickName, thisAccountLikes, accType, index, isHome }) => {
     const theme = useTheme()
@@ -19,12 +21,15 @@ export const TimelineCard = ({ avatar, name, place, caption, links, time, date, 
     const maxLength = 280
     const navigation = useNavigation()
     const [isActive, setIsActive] = useState(false)
+    const [wantLogin, setWantLogin] = useState(false)
+    const [wantText, setWantText] = useState('')
     const [isLiked, setIsLiked] = useState(false)
     const [comment, setComment] = useState('')
     const [isButtonSendActive, setIsButtonSendActive] = useState(false)
     const [images, setImages] = useState([])
     const [readMore, setReadMore] = useState(true)
     const { timelineService } = useDep()
+    const user = useSelector(state=>state.auth)
     const post = {
         caption_post: caption,
         detail_media_feed: links
@@ -72,35 +77,49 @@ export const TimelineCard = ({ avatar, name, place, caption, links, time, date, 
     }
 
     const handleLike = async () => {
-        try {
-            if (isLiked) {
-                await timelineService.doDeleteTimelineLike({
-                    "account_id": `${accId}`,
-                    "feed_id": `${feedId}`
-                })
-                setIsLiked(prevState => false)
-                setRefresh(feedId)
-            } else {
-                await timelineService.doPostTimelineLike({
-                    "account_id": `${accId}`,
-                    "feed_id": `${feedId}`
-                })
-                setIsLiked(prevState => true)
-                setRefresh(feedId)
+        if (user.accountId != 0) {
+            try {
+                if (isLiked) {
+                    await timelineService.doDeleteTimelineLike({
+                        "account_id": `${accId}`,
+                        "feed_id": `${feedId}`
+                    })
+                    setIsLiked(prevState => false)
+                    setRefresh(feedId)
+                } else {
+                    await timelineService.doPostTimelineLike({
+                        "account_id": `${accId}`,
+                        "feed_id": `${feedId}`
+                    })
+                    setIsLiked(prevState => true)
+                    setRefresh(feedId)
+                }
+            } catch (e) {
+                console.log(e);
             }
-        } catch (e) {
-            console.log(e);
+        } else {
+            handleLoginStatus()
+            setWantText(prevState=>'like this Post ?')
         }
     }
 
+    const handleLoginStatus = () => {
+        setWantLogin(prevState=>!prevState)
+    }
+
     const handleShare = async () => {
-        try {
-            await Share.share({
-                message: 'Check this post on TokTok App! https://itdp-group3-frontend-git-staging-anu-g.vercel.app/feeds'
-            })
-        } catch (err) {
-            console.log(err);
-            checkErr(err)
+        if (user.accountId != 0) {
+            try {
+                await Share.share({
+                    message: 'Check this post on TokTok App! https://itdp-group3-frontend-git-staging-anu-g.vercel.app/feeds'
+                })
+            } catch (err) {
+                console.log(err);
+                checkErr(err)
+            }   
+        } else {
+            handleLoginStatus()
+            setWantText(prevState=>'share this Post ?')
         }
     }
 
@@ -113,6 +132,7 @@ export const TimelineCard = ({ avatar, name, place, caption, links, time, date, 
     return (
         <MainContainer>
             <View style={[styles.timelineCtn, isHome && index === 0 && { marginTop: 48 }]}>
+                {wantLogin && <LoginModalComponent handleNoLogin={handleLoginStatus} text={wantText}/>}
                 <View>
                     <View style={styles.profileHd}>
                         <View style={{ flex: 1 }}>
